Datasets are now separated in different python files under the 'datasets' folder. Minor bugfix in KNN accuracy evaluator
